+++ src/components/manage/ManageQuestions.jsx
@@ -1,9 +1,10 @@
 import { async } from 'q';
 import React, { useEffect, useState } from 'react'
-import AppTable from './AppTable'
-import Searchbar from './Searchbar'
-import { getLocalUser } from '../helpers/ManageLocalStorage';
-import { getUsers } from '../middleware/api';
+import AppTable from '../table/AppTable'
+import Searchbar from '../navigation/Searchbar'
+import { getLocalUser } from '../../helpers/ManageLocalStorage';
+import { getUsers } from '../../middleware/api';
+import { getQuestions } from "../../middleware/apiQuestions"
 import { useNavigate } from 'react-router';
 
 
@@ -11,21 +12,22 @@ import { useNavigate } from 'react-router';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
-import AppModal from './AppModal';
-import CreateModal from './CreateModal';
-import DeleteModal from './DeleteModal';
+
+import CreateModal from '../modal/CreateModal';
+import QuestionsTable from '../table/QuestionsTable';
+import DeleteModal from '../modal/DeleteModal';
 
 
 
-const ManageUsers = () => {
+const ManageQuestions = () => {
 
   const navigate = useNavigate()
-  const [users, setUsers] = useState([]);
+  const [questions, setQuestions] = useState([]);
 
   useEffect(() => {
    const user = getLocalUser()
    if (user?.token){
-    loadUsers()
+    loadQuestions()
   }else{
     navigate("/auth/login")
   }
@@ -33,14 +35,14 @@ const ManageUsers = () => {
 }, []);
 
 
-const loadUsers = async ()=>{
+const loadQuestions = async ()=>{
   
   try {
-    const response = await getUsers()
+    const response = await getQuestions()
 
     if(response?.state) {
-      localStorage.setItem("users",JSON.stringify(response));
-      setUsers(response?.users)
+      localStorage.setItem("questions",JSON.stringify(response));
+      setQuestions(response?.questions)
 
       
     }else{
@@ -57,10 +59,11 @@ const loadUsers = async ()=>{
       position: "bottom-right",
       });
   }
+
 }
 
-const [modal, setModal] = useState({open: false,type: "editUser",  data: null});
-const [modalDelete, setModalDelete] = useState({open: false,type: "deleteUser",  data: null});
+const [modal, setModal] = useState({open: false,type: "editQuestion",  data: null});
+const [modalDelete, setModalDelete] = useState({open: false,type: "deleteQuestion",  data: null});
 
 const handleModal = (event) => {
   setModal(event)
@@ -70,10 +73,13 @@ const handleModalDelete = (event) => {
 }
 
 const handleRefresh = () => {
-  loadUsers()
+  loadQuestions()
 }
 
-const handleFindUser = async(query)=> {
+const handleFindQuestion = async(query)=> {
+    console.log(query)
+  /*
+  
   //event == user to find
   //en users buscar juan
   await loadUsers()
@@ -92,16 +98,20 @@ const handleFindUser = async(query)=> {
     setUsers(usersFinded)
 
     //console.log(users)
-
-  }
-
+ */
 }
+
+
 
   return (
     <div style={{width: "80%", marginLeft: "15px"}}>
-      <Searchbar type="usuario" loadAll={()=>{handleRefresh()}} findUser={ (query)=>{handleFindUser(query)}}/>
+
+      <Searchbar type="usuario" loadAll={()=>{handleRefresh()}} findUser={ (query)=>{handleFindQuestion(query)}}/>
+
       <br />
-      <AppTable usersArray={users} activeModalDelete={(event)=>{handleModalDelete(event)}} activeModal={(event)=>{handleModal(event)}}/>
+      
+      <QuestionsTable questionsArray={questions} activeModalDelete={(event)=>{handleModalDelete(event)} } activeModal={(event)=>{handleModal(event)} }/>
+
       <ToastContainer 
             position="top-center"
             autoClose={10000}
@@ -114,14 +124,16 @@ const handleFindUser = async(query)=> {
             pauseOnHover
             theme="light" />
 
-      
+    
+
       <CreateModal opened={modal?.open} activeModal={(event)=>{handleModal(event)}} refresh={()=>{handleRefresh()}} type={modal?.type} data={modal?.data}/>
 
       <DeleteModal opened={modalDelete?.open} activeModal={(event)=>{handleModalDelete(event)}} refresh={()=>{handleRefresh()}} type={modalDelete?.type} data={modalDelete?.data}/>
+
 
 
     </div>
   )
 }
 
-export default ManageUsers
+export default ManageQuestions
