+++ src/components/navigation/AppNavbar.jsx
@@ -1,7 +1,7 @@
 
 import React from 'react'
 import { Link } from 'react-router-dom'
-import Logo from './logo/Logo'
+import Logo from '../logo/Logo'
 
 
 const AppNavbar = ({user}) => {
