+++ src/store/auth/thunks.js
@@ -1,5 +1,5 @@
 //import { useDispatch } from "react-redux"
-import { registerUserWithEmailAndPassword, signInWithGoogle } from "../../firebase/providers"
+//import { registerUserWithEmailAndPassword, signInWithGoogle } from "../../firebase/providers"
 
 import {chechingCredentials,login} from "./authSlice"
 
@@ -9,23 +9,6 @@ export const checkingAuthentication = ()=>{
         dispatch(chechingCredentials())
     }
 }
-
-export const startGoogleSignIn=()=>{
-    return async (dispatch)=>{
-        dispatch(chechingCredentials())
-        const result = await signInWithGoogle()
-        dispatch(login(result))
-    }
-}
-
-export const startCreatingWithEmailAndPassword=(email,password,displayName)=>{
-
-    return async (dispatch)=>{
-        const res = await registerUserWithEmailAndPassword(email,password,displayName)
-        console.log(res)
-    }
-}
-
 
 export const saveUserLogin = (data)=>{
     return async (dispatch)=>{
